# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto APIs
  version: 1.0.5
  description: >
    **"Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!"**
    
    WASAPhoto is web application based on RESTfull architecture to allows upload your photos directly from your PC, and they will be visible to everyone following you.
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  
  contact:
    name: Andrea Pizzi
    url: https://github.com/pizzi-andrea/WASAPhoto
    email: pizzi.1995517@studenti.uniroma1.it
servers:
  - url: http://localhost:3000
    description: main server
  - url: http://localhost:4000
    description: debug server
paths:
...
  /users/{uid}/:
    summary:  users
    description: >
       Resource rappresent a single user
       
    parameters:
        - $ref: "#/components/parameters/Uid"
    
    get:
      tags:
        - user
      summary: get specific user profile
      description: >
        give user id and put user profile. User profile is object contain 
        all information on user, in particular:
          - information about user
          - stream photos updated
          - number of photo have been updated
          - number of followers
          - number of following
      
      security:
        - BearerAuth: []
        
      operationId: getUserProfile # required - used
      
      responses:
        "200":
          description: >
            get user profile corresponding to *uid*
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        
        "400":
          $ref: "#/components/responses/BadRequest"
        
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"

    put:
      tags:
        - user
      summary: assign or update username
      description: >
        assign username or update old username with new username.
        The username to set is in the body request
      
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Username"
        description: >
          new username
        required: true
      
      operationId: setMyUserName # required - used
      
      security:
        - BearerAuth: []
      
      responses:
        "204":
          description: >
            success, assigned new username
        
        "400":
          $ref: "#/components/responses/BadRequest"
          
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/banned/{bannedId}:
    summary: users banned
    description: >
      this resource describe banned user
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/BannedId"
    
    put:
      tags:
        - user
        - ban
      summary: ban user identificated by *uid* 
      description: >
        taken uid of the user who wants to ban and uid of the user to be banned, 
        bans the user(BannedID)
              
      security:
        - BearerAuth: []
      
      operationId: banUser # required - used
      
      responses:
        "201":
          description: >
            success, banned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        
        "204":
          description: >
            empty response, user just banned
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    delete:
      tags:
        - user
        - ban
      summary: unban user
      description: >
        taken uid of the user who wants unbband id of user banned and delete last one
      security:
        - BearerAuth: []
        
      operationId: unbanUser # required - used
      
      responses:
        "204":
          description: >
            success, unbanned user
        "400":
          $ref: "#/components/responses/BadRequest"
        
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
    
    get:
      tags:
        - user
        - ban
      summary: check if user had banned other user
      description: >
        taken uid of the user who wants check if banned. I if users had banned return it otherwise return empty response
      security:
        - BearerAuth: []
        
      operationId: isBanned # OK - used
      
      responses:
        "200":
          description: >
            success, user banned
          
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "204":
          description: >
            user not banned

        "400":
          $ref: "#/components/responses/BadRequest"
        
        "404":
          $ref: "#/components/responses/UserNotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/{photoId}/likes/{likeUserId}:
    
    summary: likes on photos
    description: >
      this endpoint rappresents likes putted on photo
    
    
    
    parameters:
        - $ref: "#/components/parameters/Uid"
        - $ref: "#/components/parameters/PhotoId"
        - $ref: "#/components/parameters/likeUserId"
    
    
    get:
      tags:
        - photo
        - like
      summary: check if user likeUserId has putted like on photo
      description: >
        Check if user identified by his uid has putted like on photo.
        If like has putted will return like otherwire return not found response
      
      operationId: checkLike # used
      
      security:
        - BearerAuth: []
      
      responses:
        "200":
          description: >
            success, like founded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: >
            not found, like not found
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
          
        
    
    put:
      
      tags:
        - photo
        - like
      summary: put like a photo
      description: >
        provide photo id and owner id. 
        The like is identified by the *uid* of the user who likes it.
      
      operationId: likePhoto # required - used
      
      security:
        - BearerAuth: []
      responses:
        "201":  
          description: >
            success, like added to the photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        
        "204": 
          description: >
            not content, just liked this photo
        
        "400":
          $ref: "#/components/responses/BadRequest"
        
        "404":
          description: >
            Not found, user or photo not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
      
    delete:
      tags:
        - photo
        - like
      summary: remove like a photo
      description: >
        provide photo id and owner id
        and remove like given user identificated by
        *likeUserId*
        
      operationId: unlikePhoto # required - used
      security:
        - BearerAuth: []
      responses:
        "204":  
          description: >
            success, unliked the photo
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: >
            Not found, user or photo not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{uid}/myPhotos/{photoId}/comments/{commentId}:
    summary: comment on photo
    description: >
      this resource rappresent a single comment on photo
    
    parameters:
      - $ref: "#/components/parameters/Uid"
      - $ref: "#/components/parameters/PhotoId"
      - $ref: "#/components/parameters/CommentId"
    delete:
      tags:
        - comment
        - photo
      summary: delete comment on photo
      description: >
        give user id of owner photo, photo id and remove comment on photo identified from comment id given
      
      security:
        - BearerAuth: []
        
      operationId: uncommentPhoto # required - used
      responses:
        "204":
          description: >
            success, deleted comment
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: >
            not found, photo or user not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/UnauthorizedToken"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
...
    Uid:
      name: uid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        user id
    PhotoId:
      name: photoId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
      description: >
        photo id
    CommentId:
      name: commentId
      in: path
      required: true
      description: >
        comment id
      schema:
        $ref: "#/components/schemas/Id"
    likeUserId:
      name: likeUserId
      in: path
      required: true
      description: >
        uid user put like
      schema:
        $ref: "#/components/schemas/Id"
...
    BannedId:
      name: bannedId
      in: path
      required: true
      description: >
        uid user to ban
      schema:
        $ref: "#/components/schemas/Id"
  responses:
    UnauthorizedError:
      description: >
        Access token is missing or invalid 
    UnauthorizedToken:
      description: >
        the submitted token does not have the necessary permissions to access this resource
    ServerError:
      description: >
        server internal error, for more information see log file
    BadRequest:
      description: >
        Bad request, invalid request format
    UserNotFound:
      description: >
        Not found, user not found
  securitySchemes:
    
    BearerAuth:
      description: >
        basic safety mechanism based on "bearer" mechanism. APIs that require a minimum level of security can adopt this method
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      description: >
        this object rappresent a user
      type: object
      properties:
        uid:
          $ref: "#/components/schemas/Id"
        username: 
          $ref: "#/components/schemas/Username"
      required:
        - uid
        - username
    Username:
      description: >
        username of a user
      type: string
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      example: "marioross"
    Id:
      description: >
        Identificator at 64-bit
      type: integer
      format: int64 
      readOnly: true
      example: 1
    Profile:
      description: >
        user profile rappresentation
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        stream:
          $ref: "#/components/schemas/Stream"
        follower:
          type: integer
          minimum: 0
          description: >
            number user that follow a specific user
          example: 12
        following:
          type: integer
          minimum: 0
          description: >
            numer of users following by specific user
          example: 70
    Post:
      description: >
        this object rappresent a post. A post is provided to photo and list of like and comments
        that recived  
        
      type: object
      properties:

        owner:
          $ref: "#/components/schemas/Id"
        refer:
          $ref: "#/components/schemas/Id"
        location:
          $ref: "#/components/schemas/URL"
        likes:
          description: >
            list of users who have liked it
          type: array
          items: 
            $ref: "#/components/schemas/Id"
          minItems: 0
          maxItems: 250 # da rivedere
        
        comments:
          description: >
            list of comments collected by photo
          type: array
          items:
            $ref: "#/components/schemas/Comment"
          minItems: 0
          maxItems: 250 # okok
        
        timeUpdate:
          $ref: "#/components/schemas/TimeStamp"
        
        descriptionImg:
          description: >
            image description
          type: string
          example: "a good photo!"
          minLength: 1
          maxLength: 250
          pattern: "^.*?$"
    URL: 
      description: >
        String that uniquely identifies a resource on the network
      type: string
      pattern: '^.*?$' # TODO
      minLength: 1
      maxLength: 2048
      example: "http://myExample.com::3030/5"
    Stream:
      description: >
        model of stream of photos
      type: array
      items:
        $ref: "#/components/schemas/Post"
      minItems: 0
      maxItems: 100
    TimeStamp:
      description: >
        this components describe timestamp value
        conform to [RFC3339](https://datatracker.ietf.org/doc/html/rfc3339#section-5.6) specification
      type: string
      format: data-time
      pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      example: "2017-07-21T17:32:28Z"
      minLength: 20
      maxLength: 21
      readOnly: true
...
    Comment:
      description: > 
        this object rappresent a comment on a photo.
      type: object
      properties:
        commentId:
          $ref: "#/components/schemas/Id"
        author:
          $ref: "#/components/schemas/User"
  
        text:
          description: >
            comment text encoded in UNICODE format
          type: string
          minLength: 1
          maxLength: 250
          pattern: "^.*$"
          example: 
            "😀 i like you photo! 😀"
        timeStamp:
          $ref: "#/components/schemas/TimeStamp"
      required:
        - author
        - text
        - timeStamp
...